refactor(Card): 메서드를 블랙잭에 종속적이지 않도록 수정
