+++ src/main/java/blackjack/domain/GamePoint.java
@@ -1,10 +1,17 @@
-package blackjack.domain.card;
+package blackjack.domain;
+
+import blackjack.domain.card.Card;
 
 import java.util.List;
 
-public class GamePoint {
+public final class GamePoint {
 
     private static final int BUST = 0;
+    public static final int ACE = 1;
+    public static final int ACE_UPPER = 11;
+    public static final int FACE_CARD = 10;
+    public static final int BLACK_JACK = 21;
+
     private final int gamePoint;
 
     public GamePoint(final List<Card> cards) {
@@ -16,10 +23,6 @@ public class GamePoint {
         this.gamePoint = calculateBust(point);
     }
 
-    private boolean isBust(final int point) {
-        return calculateBust(point) == BUST;
-    }
-
     private int calculateMaxPoint(final List<Card> cards) {
         int point = 0;
         for (Card card : cards) {
@@ -28,10 +31,18 @@ public class GamePoint {
         return point;
     }
 
+    private boolean containAceInCards(List<Card> cards) {
+        return getCountOfAce(cards) != 0;
+    }
+
+    private boolean isBust(final int point) {
+        return calculateBust(point) == BUST;
+    }
+
     private int getPoint(final int point, final int aceCount) {
         int optimizedPoint = point;
         int remainAce = aceCount;
-        while (optimizedPoint > 21 && remainAce > 0) {
+        while (optimizedPoint > BLACK_JACK && remainAce > 0) {
             optimizedPoint -= 10;
             remainAce -= 1;
         }
@@ -39,32 +50,25 @@ public class GamePoint {
         return calculateBust(optimizedPoint);
     }
 
-    private boolean containAceInCards(List<Card> cards) {
-        if (getCountOfAce(cards) != 0) {
-            return true;
-        }
-        return false;
-    }
-
-    private static int getCountOfAce(final List<Card> cards) {
+    private int getCountOfAce(final List<Card> cards) {
         return (int) cards.stream()
-                .filter((card) -> card.isAce())
+                .filter(card -> card.isSameAs(ACE))
                 .count();
     }
 
     private int calculateBust(final int point) {
-        if (point > 21) {
+        if (point > BLACK_JACK) {
             return BUST;
         }
         return point;
     }
 
     private int transform(Card card) {
-        if (card.isAce()) {
-            return 11;
+        if (card.isSameAs(ACE)) {
+            return ACE_UPPER;
         }
-        if (card.isOverTen()) {
-            return 10;
+        if (card.isOver(FACE_CARD)) {
+            return FACE_CARD;
         }
         return card.getCardNumberValue();
     }
