+++ src/main/java/blackjack/domain/Users.java
@@ -1,7 +1,6 @@
 package blackjack.domain;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.GamePoint;
 import blackjack.domain.user.Name;
 import blackjack.domain.user.User;
